requirements.txt
beautifulsoup4==4.10.0
Pillow==8.4.0
requests==2.26.0
rich==10.9.0
